--- a/src/manifold.jl
+++ b/src/manifold.jl
@@ -1,13 +1,13 @@
-Base.@pure function determine_chunksize(u, alg::DiffEqBase.DEAlgorithm)
-    determine_chunksize(u, get_chunksize(alg))
-end
-Base.@pure function determine_chunksize(u, CS)
-    if CS != 0
-        return CS
-    else
-        return ForwardDiff.pickchunksize(length(u))
-    end
-end
+# Base.@pure function determine_chunksize(u, alg::DiffEqBase.DEAlgorithm)
+#     determine_chunksize(u, get_chunksize(alg))
+# end
+# Base.@pure function determine_chunksize(u, CS)
+#     if CS != 0
+#         return CS
+#     else
+#         return ForwardDiff.pickchunksize(length(u))
+#     end
+# end
 
 struct NLSOLVEJL_SETUP{CS, AD} end
 Base.@pure function NLSOLVEJL_SETUP(; chunk_size = 0, autodiff = true)
@@ -21,13 +21,16 @@ function (p::NLSOLVEJL_SETUP{CS, AD})(::Type{Val{:init}}, f, u0_prototype) where
 end
 
 # wrapper for non-autonomous functions
-mutable struct NonAutonomousFunction{F, autonomous}
+mutable struct NonAutonomousFunction{iip, F, autonomous, T}
     f::F
-    t::Any
-    p::Any
+    t::T
 end
-(p::NonAutonomousFunction{F, true})(res, u) where {F} = p.f(res, u, p.p)
-(p::NonAutonomousFunction{F, false})(res, u) where {F} = p.f(res, u, p.p, p.t)
+(f::NonAutonomousFunction{true, F, true})(res, u, p) where {F} = f.f(res, u, p)
+(f::NonAutonomousFunction{true, F, false})(res, u, p) where {F} = f.f(res, u, p, p.t)
+(f::NonAutonomousFunction{false, F, true})(u, p) where {F} = f.f(u, p)
+(f::NonAutonomousFunction{false, F, false})(u, p) where {F} = f.f(u, p, p.t)
+
+SciMLBase.isinplace(::NonAutonomousFunction{iip}) where {iip} = iip
 
 """
 ```julia
@@ -81,50 +84,55 @@ Ernst Hairer, Christian Lubich, Gerhard Wanner. Geometric Numerical Integration:
 Structure-Preserving Algorithms for Ordinary Differential Equations. Berlin ;
 New York :Springer, 2002.
 """
-mutable struct ManifoldProjection{autonomous, F, NL, O}
+mutable struct ManifoldProjection{iip, autonomous, F, NL, NO}
     g::F
     nl_rhs::Any
-    nlsolve::NL
-    nlopts::O
+    nlsolver::NL
+    nlopts::NO
 
-    function ManifoldProjection{autonomous}(g, nlsolve, nlopts) where {autonomous}
-        # replace residual function if it is time-dependent
-        # since NLsolve only accepts functions with two arguments
-        _g = NonAutonomousFunction{typeof(g), autonomous}(g, 0, 0)
-        new{autonomous, typeof(_g), typeof(nlsolve), typeof(nlopts)}(_g, _g, nlsolve,
-            nlopts)
+    # TODO: Caching for NonlinearSolve solvers
+    function ManifoldProjection{iip, autonomous}(g, nlsolve, nlopts) where {iip, autonomous}
+        replace residual function if it is time-dependent
+        _g = NonAutonomousFunction{iip, typeof(g), autonomous}(g, 0)
+        return new{iip, autonomous, typeof(_g), typeof(nlsolve), typeof(nlopts)}(_g, _g,
+            nlsolve, nlopts)
     end
 end
 
-# Now make `affect!` for this:
-function (p::ManifoldProjection{autonomous, NL})(integrator) where {autonomous, NL}
-    # update current time if residual function is time-dependent
-    if !autonomous
-        p.g.t = integrator.t
+# # Now make `affect!` for this:
+# function (p::ManifoldProjection{autonomous, NL})(integrator) where {autonomous, NL}
+#     # update current time if residual function is time-dependent
+#     if !autonomous
+#         p.g.t = integrator.t
+#     end
+#     p.g.p = integrator.p
+
+#     integrator.u .= p.nlsolve(p.nl_rhs, integrator.u; p.nlopts...)
+# end
+
+# function Manifold_initialize(cb, u::Number, t, integrator)
+#     cb.affect!.nl_rhs = cb.affect!.nlsolve(Val{:init}, cb.affect!.g, [u])
+#     u_modified!(integrator, false)
+# end
+
+# function Manifold_initialize(cb, u, t, integrator)
+#     cb.affect!.nl_rhs = cb.affect!.nlsolve(Val{:init}, cb.affect!.g, u)
+#     u_modified!(integrator, false)
+# end
+
+function ManifoldProjection(g; nlsolve = nothing, save = true, isinplace = Val(true),
+        autonomous = nothing, nlopts = (;))
+    if autonomous === nothing
+        if SciMLBase._unwrap_val(isinplace)
+            autonomous = maximum(SciMLBase.numargs(g)) == 3
+        else
+            autonomous = maximum(SciMLBase.numargs(g)) == 2
+        end
     end
-    p.g.p = integrator.p
-
-    integrator.u .= p.nlsolve(p.nl_rhs, integrator.u; p.nlopts...)
-end
-
-function Manifold_initialize(cb, u::Number, t, integrator)
-    cb.affect!.nl_rhs = cb.affect!.nlsolve(Val{:init}, cb.affect!.g, [u])
-    u_modified!(integrator, false)
-end
-
-function Manifold_initialize(cb, u, t, integrator)
-    cb.affect!.nl_rhs = cb.affect!.nlsolve(Val{:init}, cb.affect!.g, u)
-    u_modified!(integrator, false)
-end
-
-function ManifoldProjection(g; nlsolve = NLSOLVEJL_SETUP(), save = true,
-        autonomous = maximum(SciMLBase.numargs(g)) == 3,
-        nlopts = Dict{Symbol, Any}())
     affect! = ManifoldProjection{autonomous}(g, nlsolve, nlopts)
     condition = (u, t, integrator) -> true
     save_positions = (false, save)
-    DiscreteCallback(condition, affect!;
-        initialize = Manifold_initialize,
+    return DiscreteCallback(condition, affect!; initialize = Manifold_initialize,
         save_positions = save_positions)
 end