--- a/test/manifold_tests.jl
+++ b/test/manifold_tests.jl
@@ -10,19 +10,21 @@ prob = ODEProblem(f, u0, (0.0, 100.0))
 function g(resid, u, p)
     resid[1] = u[2]^2 + u[1]^2 - 2
     resid[2] = u[3]^2 + u[4]^2 - 2
-    resid[3] = 0
-    resid[4] = 0
 end
 
 g_t(resid, u, p, t) = g(resid, u, p)
 
-isautonomous(p::ManifoldProjection{autonomous, NL}) where {autonomous, NL} = autonomous
+function isautonomous(::ManifoldProjection{
+        iip, nlls, autonomous, NL}) where {iip, nlls, autonomous, NL}
+    return autonomous
+end
 
 sol = solve(prob, Vern7())
 @test !(sol[end][1]^2 + sol[end][2]^2 ≈ 2)
 
 # autodiff=true
-cb = ManifoldProjection(g)
+@inferred ManifoldProjection(g; autonomous = Val(true), resid_prototype = zeros(2))
+cb = ManifoldProjection(g; resid_prototype = zeros(2))
 @test isautonomous(cb.affect!)
 solve(prob, Vern7(), callback = cb)
 @time sol = solve(prob, Vern7(), callback = cb)