--- a/test/manifold_tests.jl
+++ b/test/manifold_tests.jl
@@ -1,4 +1,4 @@
-using OrdinaryDiffEq, Test, DiffEqBase, DiffEqCallbacks, RecursiveArrayTools
+using OrdinaryDiffEq, Test, DiffEqBase, DiffEqCallbacks, RecursiveArrayTools, NonlinearSolve
 
 u0 = ones(2, 2)
 f = function (du, u, p, t)
@@ -23,29 +23,29 @@ sol = solve(prob, Vern7())
 @test !(sol[end][1]^2 + sol[end][2]^2 ≈ 2)
 
 # autodiff=true
-@inferred ManifoldProjection(g; autonomous = Val(true), resid_prototype = zeros(2))
+@inferred ManifoldProjection(g; autonomous = Val(false), resid_prototype = zeros(2))
 cb = ManifoldProjection(g; resid_prototype = zeros(2))
 @test isautonomous(cb.affect!)
 solve(prob, Vern7(), callback = cb)
 @time sol = solve(prob, Vern7(), callback = cb)
 @test sol[end][1]^2 + sol[end][2]^2 ≈ 2
 
-cb_t = ManifoldProjection(g_t)
+cb_t = ManifoldProjection(g_t; resid_prototype = zeros(2))
 @test !isautonomous(cb_t.affect!)
 solve(prob, Vern7(), callback = cb_t)
 @time sol_t = solve(prob, Vern7(), callback = cb_t)
 @test sol_t.u == sol.u && sol_t.t == sol.t
 
 # autodiff=false
-cb_false = ManifoldProjection(g,
-    nlsolve = DiffEqCallbacks.NLSOLVEJL_SETUP(autodiff = false))
+cb_false = ManifoldProjection(
+    g; nlsolve = GaussNewton(; autodiff = AutoFiniteDiff()), resid_prototype = zeros(2))
 @test isautonomous(cb_false.affect!)
 solve(prob, Vern7(), callback = cb_false)
 sol = solve(prob, Vern7(), callback = cb_false)
 @test sol[end][1]^2 + sol[end][2]^2 ≈ 2
 
 cb_t_false = ManifoldProjection(g_t,
-    nlsolve = DiffEqCallbacks.NLSOLVEJL_SETUP(autodiff = false))
+    nlsolve = GaussNewton(; autodiff = AutoFiniteDiff()), resid_prototype = zeros(2))
 @test !isautonomous(cb_t_false.affect!)
 solve(prob, Vern7(), callback = cb_t_false)
 sol_t = solve(prob, Vern7(), callback = cb_t_false)
@@ -61,10 +61,27 @@ sol = solve(prob, Vern7(), callback = cb)
 sol = solve(prob, Vern7(), callback = cb_t)
 @test sol[end][1]^2 + sol[end][2]^2 ≈ 2
 
-# does not work since Calculus.jl (on which NLsolve.jl depends)
-# implements only Jacobians of vectors
 sol = solve(prob, Vern7(), callback = cb_false)
-sol[end][1]^2 + sol[end][2]^2 ≈ 2
+@test sol[end][1]^2 + sol[end][2]^2 ≈ 2
 
 sol = solve(prob, Vern7(), callback = cb_t_false)
-sol[end][1]^2 + sol[end][2]^2 ≈ 2
+@test sol[end][1]^2 + sol[end][2]^2 ≈ 2
+
+# Test termination if cannot project to manifold
+function g_unsat(resid, u, p)
+    resid[1] = u[2]^2 + u[1]^2 - 1000
+    resid[2] = u[2]^2 + u[1]^2 - 20
+end
+
+cb_unsat = ManifoldProjection(g_unsat; resid_prototype = zeros(2))
+sol = solve(prob, Vern7(), callback = cb_unsat)
+@test !SciMLBase.successful_retcode(sol)
+@test last(sol.t) != 100.0
+
+# Tests for OOP Manifold Projection
+function g_oop(u, p)
+    return [u[2]^2 + u[1]^2 - 2
+            u[3]^2 + u[4]^2 - 2]
+end
+
+g_t(resid, u, p, t) = g(resid, u, p)