@avik-pal 2024-02-20
Also ideally needs https://github.com/SciML/NonlinearSolve.jl/pull/382